+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 This project changes will be shown here.
 
+## Unreleased - 2021-08-09
+### Changed
+- Added new itemCategory in RenContractComponent for Issue 1481
+
 ## [v5.15.4] - 2021-08-02
 ### Changed
 - Changed the way information is fetched to generate sales orders from contracts to try and get it always sorted the same way.